Update linting Github Workflow file
Removed the concurrency settings from the linting Github Workflow file `.github/workflows/linting.yml`. The `concurrency` property was responsible for organizing the jobs in queue but it was unnecessary in this case because each linting job should run independently and there's no need to cancel any in-progress job. Hence, by removing these four lines of code, the structure of the workflow file is simplified and the risk of unintentionally cancelled jobs is also eliminated.