Refactor BannerRegister and add Avatar feature
Renamed struct in banner.rs from BannerRegister to RegisterLocalisedBanner for consistency.

In addition, implemented a new feature in avatar.rs, allowing to show the profile of a user. Relevant structs and JSON language files were added to accommodate this feature. This enhances the user's ability to interact and gain information about other users.