Update readme.md with project status changes
Various changes have been made in project status. Code cleanup task remarked as never-ending, avatar feature is now in progress, and DB change to PostgreSQL highlighted in readme.md.

The Dockerfile is now using a Debian base image. Efforts on better error handling are ongoing.

Localisation for response and command are progressing but help is needed to proofread the JSON file and continue the missing parts.

Despite this, continuation for renaming function, structures, command names is still paused. The need to add docs to every public function is also recognized.