Update Docker configuration to use stable image
Replaced the Docker build context with a stable, pre-built image in `compose-default.yml`. The previous build context required a local Dockerfile, which might not always represent the latest or most stable version of the service. By using directly the latest stable version of 'kasuki' image, it guarantees that the service will run against the most well-tested, up-to-date version of the application.