Refactor variable naming and iteration method.
Renamed 'langages' to 'available_langages' and 'langs' to 'langages' in lang.rs for clarity and consistency. Changed iteration methods in lang.rs, info.rs, credit.rs and banner.rs to directly access the values of HashMaps, excluding unused keys, for a more efficient and cleaner code. Simplified the assignment of 'state' in `module_activation.rs` to reduce unnecessary clone operation.