Update 'last_page' value and json query in Anilist module
The past default 'last_page' value on the Anilist endpoint has been changed from 1444 to 1567 to reflect updates in the Anilist database. Also, in the json request payload, the query is updated from querying "anime_page" to "manga_page" for a more appropriate result. This helps in fetching the correct data for manga instead of animes.