Refactor Github Actions workflows to include manual triggers
The GitHub Actions workflows for linting, testing, and security audit have been updated to include the 'workflow_dispatch' event. This change allows for manual triggering of these workflows, in addition to the current triggers on push and pull_request events. This increases flexibility with executing these automations, especially when immediate action is necessary without a code modification event.