Add avatar command and language file
This commit introduces a new 'avatar' command that displays a user's avatar. Additionally, a new 'avatar.json' language file has been created. Previously we had 'profile' command which showed the profile of the user, now we changed it to ensure our command shows the avatar of the user instead. This enhances the bot's utility and makes user avatars more accessible to other users in the server.