Update status and remove Codacy Badge in readme.md
This commit updates the status of the 'discord bot linked to anilist' from 'not working' to 'working again' in the acknowledgments section of the readme.md file. It also removes the Codacy badge for code health. The badge is removed because it wasn't providing significant value. The status update provides up-to-date information about the functionality of the bot.