Refactor color usage in codebase
Updated codebase to use a constant value for color (FABLED_PINK) instead of defining it in multiple places. A constant COLOR has been imported from the crate::constant module at the start of each file and replaced the local definitions of color or Colour::FABLED_PINK. This change will make it easier to manage and update the color as it is now defined in a single place.