Remove security audit workflow
The security audit workflow was removed as it was no longer needed in the current project. This workflow was initially added for regularly checking potential security vulnerabilities using the "rustsec/audit-check" action. However, improvements in our codebase structure and the integration of a different security solution has made this workflow redundant. Hence, based on the recent technical discussions the team decided to remove this workflow from our GitHub actions.