Add logging functionality and update environment configs
Implemented a simple logger in `src/logger.rs` for easy debugging and traceability of any issues. A new environment variable `LOG` is added to configure the verbosity of logs in `compose-default.yml` and `.env-default`. Logger is initialized in the main function and `LOG` level is read from the environments. Updated `Cargo.toml` to include the necessary libraries. The logging level can be set to "trace", "debug", "info", "warn" or "error". This commit introduced a more enhanced way of monitoring the activities of the bot, making debugging easier.