Refactor minimal anime search function
Changes have been made to the way minimal anime details are fetched. Rather than searching by a string as was previously done, the function now fetches anime details using an id number. Moreover, error handling has been improved in the event that an anime id is not provided or if it's in an incorrect format. The primary logic behind this modification is to ensure accuracy of the data retrieved because ids are much more precise compared to strings.