Update variable in JSON request for fetching anime
The JSON request responsible for fetching minimal anime details has been updated. Previously, the variable "name" was being used for matching anime. Now, "id" is being used. This change was made as it significantly increases the accuracy of the anime matching process, aligning the variable closer with the data's unique identifiers.