Remove redundant code and clean documentation
The superfluous code and function in `data.rs` was removed to enhance the maintainability and readability of the application. This function, `set_data_activity_sqlite`, was redundant and provided no added functionality to the application. Its removal simplifies the code structure and improves performance.

Moreover, a blank line in `readme.md` which served no purpose was removed for documentation cleanliness.