Add .gitignore rules for IDE and log files
This commit adds numerous files to the .gitignore, specifically files created by the IDE and log files. These files are typically user-specific or dynamically changing, and are not necessary for the overall functionality of the project. Including them in the .gitignore prevents these files from being committed to the git repository, keeping the repository cleaner and more focused on the necessary project files only.