Update serenity library and rust version in Dockerfile
Serenity library is updated to version 0.12.0 in 'Cargo.toml' to incorporate new features and improvements from the library. This update might fix some known bugs and introduce new features that can be beneficial for our solution. Also, updated the rust version to 1.74-buster in 'Dockerfile' which brings several new language features and bug fixes aiming to improve the runtime efficiency and stability of our application.