Add module interaction and localization support
Introduced new functionality to activate and deactivate a module in reaction to user interaction. Updated 'lang' command in lang.json to make the 'lang' option required and to reflect changes in lang.rs. Created new commands in module_activation.json to allow users to choose a module to turn on or off. New files module.rs and lang.rs were added to src/command_run/general module, implementing the new functionality using the newly written json files. Localizations for this new functionality were added in the relevant json files. SQL is used to fetch the current settings and update them as the user interacts with the bot.