Refactored the autocomplete feature for anime search
This commit refactors the autocomplete feature to deliver a streamlined search result for anime, manga and light-novel. The update simplifies certain functions, removing the need for superfluous parameters and replacing them with a constant, thereby optimizing the autocomplete count. It also takes advantage of `tracing` feature over `log` to improve efficiency and readability of debugging logs. Lastly, unused imports were removed for cleaner code. This refactoring leads to a more readable, maintainable, and performant code.